// src/BannerAPI/types/banner/CourseSearchResult.ts
// Generated by https://transform.tools/json-to-typescript

export interface BannerCourseSearchResult {
  success: boolean;
  totalCount: number;
  data: BannerCourseSearchData[];
  pageOffset: number;
  pageMaxSize: number;
  sectionsFetchedCount: number;
  pathMode: string;
  searchResultsConfigs: BannerSearchResultsConfig[];
  ztcEncodedImage: string;
}

export interface BannerCourseSearchData {
  id: number;
  term: string;
  termDesc: string;
  courseReferenceNumber: string;
  partOfTerm: string;
  courseNumber: string;
  subject: string;
  subjectDescription: string;
  sequenceNumber: string;
  campusDescription: string;
  scheduleTypeDescription: string;
  courseTitle: string;
  creditHours: number;
  maximumEnrollment: number;
  enrollment: number;
  seatsAvailable: number;
  waitCapacity: number;
  waitCount: number;
  waitAvailable: number;
  crossList?: string;
  crossListCapacity?: number;
  crossListCount?: number;
  crossListAvailable?: number;
  creditHourHigh: number;
  creditHourLow: number;
  creditHourIndicator: string;
  openSection: boolean;
  linkIdentifier: string;
  isSectionLinked: boolean;
  subjectCourse: string;
  faculty: BannerFaculty[];
  meetingsFaculty: BannerMeetingsFaculty[];
  reservedSeatSummary?: BannerReservedSeatSummary;
  sectionAttributes: BannerSectionAttribute[];
  instructionalMethod?: string;
  instructionalMethodDescription?: string;
}

export interface BannerFaculty {
  bannerId: string;
  category: string;
  class: string;
  courseReferenceNumber: string;
  displayName: string;
  emailAddress: string;
  primaryIndicator: boolean;
  term: string;
}

export interface BannerMeetingsFaculty {
  category: string;
  class: string;
  courseReferenceNumber: string;
  faculty: unknown[];
  meetingTime: BannerMeetingTime;
  term: string;
}

export interface BannerMeetingTime {
  beginTime: string;
  building: string;
  buildingDescription: string;
  campus: string;
  campusDescription: string;
  category: string;
  class: string;
  courseReferenceNumber: string;
  creditHourSession: number;
  endDate: string;
  endTime: string;
  friday: boolean;
  hoursWeek: number;
  meetingScheduleType: string;
  meetingType: string;
  meetingTypeDescription: string;
  monday: boolean;
  room: string;
  saturday: boolean;
  startDate: string;
  sunday: boolean;
  term: string;
  thursday: boolean;
  tuesday: boolean;
  wednesday: boolean;
}

export interface BannerReservedSeatSummary {
  class: string;
  courseReferenceNumber: string;
  maximumEnrollmentReserved: number;
  seatsAvailableReserved: number;
  seatsAvailableUnreserved: number;
  termCode: string;
  waitAvailableReserved: number;
  waitAvailableUnreserved: number;
  waitCapacityReserved: number;
  waitCapacityUnreserved: number;
}

export interface BannerSectionAttribute {
  class: string;
  code: string;
  courseReferenceNumber: string;
  description: string;
  isZTCAttribute: boolean;
  termCode: string;
}

export interface BannerSearchResultsConfig {
  config: string;
  display: string;
  title: string;
  required: boolean;
  width: string;
}